Update UML diagram for Model to include refactor from detail to remark Moved Name and Remark classes into their own packages to maintain OOP practices
