+++ src/main/java/seedu/partyplanet/logic/parser/EAddCommandParser.java
@@ -11,8 +11,8 @@ import seedu.partyplanet.logic.commands.EAddCommand;
 import seedu.partyplanet.logic.parser.exceptions.ParseException;
 import seedu.partyplanet.model.event.Event;
 import seedu.partyplanet.model.event.EventDate;
-import seedu.partyplanet.model.person.Name;
-import seedu.partyplanet.model.person.Remark;
+import seedu.partyplanet.model.name.Name;
+import seedu.partyplanet.model.remark.Remark;
 
 /**
  * Parses input arguments and creates a new EAddCommand object
