+++ src/main/java/seedu/partyplanet/storage/JsonAdaptedEvent.java
@@ -9,8 +9,8 @@ import seedu.partyplanet.commons.exceptions.IllegalValueException;
 import seedu.partyplanet.logic.parser.exceptions.ParseException;
 import seedu.partyplanet.model.event.Event;
 import seedu.partyplanet.model.event.EventDate;
-import seedu.partyplanet.model.person.Name;
-import seedu.partyplanet.model.person.Remark;
+import seedu.partyplanet.model.name.Name;
+import seedu.partyplanet.model.remark.Remark;
 
 /**
  * Jackson-friendly version of {@link Event}.
