+++ src/test/java/seedu/partyplanet/logic/parser/EditCommandParserTest.java
@@ -39,7 +39,7 @@ import seedu.partyplanet.logic.commands.EditFieldCommand;
 import seedu.partyplanet.logic.commands.EditFieldCommand.EditPersonDescriptor;
 import seedu.partyplanet.model.person.Address;
 import seedu.partyplanet.model.person.Email;
-import seedu.partyplanet.model.person.Name;
+import seedu.partyplanet.model.name.Name;
 import seedu.partyplanet.model.person.Phone;
 import seedu.partyplanet.model.tag.Tag;
 import seedu.partyplanet.testutil.EditPersonDescriptorBuilder;
