+++ src/test/java/seedu/partyplanet/logic/parser/ParserUtilTest.java
@@ -16,7 +16,7 @@ import org.junit.jupiter.api.Test;
 import seedu.partyplanet.logic.parser.exceptions.ParseException;
 import seedu.partyplanet.model.person.Address;
 import seedu.partyplanet.model.person.Email;
-import seedu.partyplanet.model.person.Name;
+import seedu.partyplanet.model.name.Name;
 import seedu.partyplanet.model.person.Phone;
 import seedu.partyplanet.model.tag.Tag;
