+++ src/test/java/seedu/partyplanet/storage/JsonAdaptedPersonTest.java
@@ -14,9 +14,9 @@ import org.junit.jupiter.api.Test;
 import seedu.partyplanet.commons.exceptions.IllegalValueException;
 import seedu.partyplanet.model.person.Address;
 import seedu.partyplanet.model.person.Email;
-import seedu.partyplanet.model.person.Name;
+import seedu.partyplanet.model.name.Name;
 import seedu.partyplanet.model.person.Phone;
-import seedu.partyplanet.model.person.Remark;
+import seedu.partyplanet.model.remark.Remark;
 
 public class JsonAdaptedPersonTest {
     private static final String INVALID_NAME = "R@chel";
