+++ src/test/java/seedu/partyplanet/testutil/EventBuilder.java
@@ -2,8 +2,8 @@ package seedu.partyplanet.testutil;
 
 import seedu.partyplanet.model.event.Event;
 import seedu.partyplanet.model.event.EventDate;
-import seedu.partyplanet.model.person.Name;
-import seedu.partyplanet.model.person.Remark;
+import seedu.partyplanet.model.name.Name;
+import seedu.partyplanet.model.remark.Remark;
 
 /**
  * A utility class to help with building Person objects.
